+++ test/Spec.hs
@@ -7,6 +7,7 @@ import qualified File.MapIOSpec
 import qualified Game.GridUtilsSpec
 import qualified Game.LogicSpec
 import qualified Game.StateSpec
+import qualified UI.DrawSpec
 
 main :: IO ()
 main = hspec $ do
@@ -14,3 +15,4 @@ main = hspec $ do
   describe "Game.State"     Game.StateSpec.spec
   describe "Game.Logic"     Game.LogicSpec.spec
   describe "File.MapIO"     File.MapIOSpec.spec
+  describe "UI.Draw"        UI.DrawSpec.spec